Update i1118_736.js
Tested shoulder buttons and this works. 

Triggers for xbox seem impossible with HID - I'm not sure how they worked in previous app versions, but per the Xinput HID docs, they interact as a single input ("Z") and act as two axis values for a third joystick. It would be impossible to tell if one is held down unless the other is completely released.